+++ compiler-core/src/erlang/tests.rs
@@ -22,7 +22,6 @@ mod patterns;
 mod pipes;
 mod records;
 mod reserved;
-mod statement_if;
 mod strings;
 mod todo;
 mod use_;
@@ -565,5 +564,22 @@ pub type Box {
 // https://github.com/gleam-lang/gleam/issues/2156
 #[test]
 fn dynamic() {
-    assert_erl!("pub external type Dynamic")
+    assert_erl!("pub type Dynamic")
+}
+
+// https://github.com/gleam-lang/gleam/issues/2166
+#[test]
+fn inline_const_pattern_option() {
+    assert_erl!(
+        "pub fn main() {
+            let fifteen = 15
+            let x = <<5:size(sixteen)>>
+            case x {
+              <<5:size(sixteen)>> -> <<5:size(sixteen)>>
+              <<6:size(fifteen)>> -> <<5:size(fifteen)>>
+            }
+          }
+          
+          pub const sixteen = 16"
+    )
 }