+++ compiler-core/src/javascript/tests/generics.rs
@@ -36,20 +36,6 @@ fn tuple_generics_typescript() {
 }
 
 #[test]
-fn externals_generics_typescript() {
-    assert_ts_def!(
-        r#"pub external type Queue(a)
-
-pub external fn new() -> Queue(a) = "queue" "new"
-
-pub external fn length(Queue(a)) -> Int = "queue" "len"
-
-pub external fn push(Queue(a), a) -> Queue(a) = "queue" "in"
-"#,
-    );
-}
-
-#[test]
 fn result_typescript() {
     assert_ts_def!(
         r#"pub fn map(result, fun) {
@@ -64,7 +50,7 @@ fn result_typescript() {
 #[test]
 fn task_typescript() {
     assert_ts_def!(
-        r#"pub external type Promise(value)
+        r#"pub type Promise(value)
     pub type Task(a) = fn() -> Promise(a)"#,
     );
 }