+++ compiler-core/src/language_server/engine.rs
@@ -105,8 +105,7 @@ where
         self.progress_reporter.compilation_finished();
 
         let modules = result?;
-        self.modules_compiled_since_last_feedback
-            .extend(modules.into_iter());
+        self.modules_compiled_since_last_feedback.extend(modules);
 
         Ok(())
     }
@@ -204,11 +203,9 @@ where
                     }
                 }
 
-                Located::ModuleStatement(
-                    Definition::ExternalFunction(_)
-                    | Definition::TypeAlias(_)
-                    | Definition::CustomType(_),
-                ) => Some(this.completion_types(module)),
+                Located::ModuleStatement(Definition::TypeAlias(_) | Definition::CustomType(_)) => {
+                    Some(this.completion_types(module))
+                }
 
                 Located::ModuleStatement(Definition::Import(_) | Definition::ModuleConstant(_)) => {
                     None