+++ compiler-core/src/type_/pattern.rs
@@ -228,7 +228,7 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
             }
 
             Pattern::VarUsage { name, location, .. } => {
-                let ValueConstructor { type_, .. } = self
+                let vc = self
                     .environment
                     .get_variable(&name)
                     .cloned()
@@ -238,14 +238,15 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                         variables: self.environment.local_value_names(),
                     })?;
                 self.environment.increment_usage(&name);
-                let typ = self
-                    .environment
-                    .instantiate(type_, &mut hashmap![], self.hydrator);
+                let typ =
+                    self.environment
+                        .instantiate(vc.type_.clone(), &mut hashmap![], self.hydrator);
                 unify(int(), typ.clone()).map_err(|e| convert_unify_error(e, location))?;
 
                 Ok(Pattern::VarUsage {
                     name,
                     location,
+                    constructor: Some(vc),
                     type_: typ,
                 })
             }